Fix node-sass watcher
Currently node-sass watcher is used to compile sass files when some changes occurs. The problem is when file A is imported in file B. If some change occurs in file A,  file A will be successfully compiled by `node-sass` but file B will not be updated. So when `livesync command` is executed, changes are not reflected in the app's UI.

This PR removes `--watch` option from `node-sass` compiler and introduces chokidar watcher.

Also fixes this https://github.com/NativeScript/nativescript-dev-sass/issues/70